requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
croniter==5.0.1
discord-typings==0.9.0
discord.py==2.4.0
distro==1.9.0
emoji==2.14.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
multidict==6.1.0
openai==1.54.5
propcache==0.2.0
py-cord==2.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
tomli==2.1.0
tqdm==4.67.0
typing_extensions==4.12.2
yarl==1.17.2
